Combine cognito routes in one file. Rename func and add comments
