+++ identity-service/src/cognitoFlowMiddleware.js
@@ -1,6 +1,11 @@
 const { sendResponse, errorResponseBadRequest } = require('./apiHelpers')
 const { isWebhookValid } = require('./utils/cognitoHelpers')
 
+// set a maximum allowed time drift between the flow completion and
+// the webhook transmission by cognito
+// this is so that past webhooks have a limited time window to be sent
+// https://docs.cognitohq.com/guides#receiving-webhooks
+// cognito should send the webhook soon after a user completes the flow
 const MAX_TIME_DRIFT_MILLISECONDS = 15 * 60 * 1000 // fifteen minutes
 
 /**
